% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_from_bam.R
\name{calc_string_entropy_k_mer}
\alias{calc_string_entropy_k_mer}
\title{Title}
\usage{
calc_string_entropy_k_mer(s, k = 2, alphabet = c("A", "C", "G", "T", "N"))

calc_string_entropy_k_mer(s, k = 2, alphabet = c("A", "C", "G", "T", "N"))
}
\arguments{
\item{s}{string}

\item{k}{kmer size}

\item{alphabet}{possible readings}
}
\value{


shannon entropy
}
\description{
Title

Title
}
