% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_beta.R
\name{calculate_beta_factor}
\alias{calculate_beta_factor}
\title{Calculate beta-factor}
\usage{
calculate_beta_factor(
  bam_file_path,
  factor,
  reference_path,
  bed_file = NULL,
  mm_rate_max = NULL
)
}
\arguments{
\item{bam_file_path}{Path to .BAM-file}

\item{factor}{mismatch to match ratio in training data}

\item{reference_path}{Path to reference genome e.g. FASTA-file.}

\item{bed_file}{path to bed file with relevant regions}

\item{mm_rate_max}{maximum error rate for beta calculation}
}
\value{
\code{\link[=data.frame]{data.frame()}}. Each line describes a position in a read.
}
\description{
This functions calculates the beta factor based on a BAM-file.
}
