% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_mutations.R
\name{call_mutations}
\alias{call_mutations}
\title{Call mutations from read positions}
\usage{
call_mutations(
  mutations_df,
  read_positions_df,
  model,
  beta,
  alpha = 0.05,
  use_turboem = TRUE,
  calculate_confidence_interval = FALSE
)
}
\arguments{
\item{mutations_df}{A \code{\link[=data.frame]{data.frame()}} with candidate mutations (SNVs) (chromosome, positions, reference and alternative)}

\item{read_positions_df}{A \code{\link[=data.frame]{data.frame()}} with read-positions. See \code{\link[=get_read_positions_from_BAM]{get_read_positions_from_BAM()}}}

\item{model}{A dreams model. See \code{\link[=train_dreams_model]{train_dreams_model()}}.}

\item{beta}{Down sampling parameter from (TODO: Link)}

\item{alpha}{Alpha-level used for testing and confidence intervals. Default is 0.05.}

\item{use_turboem}{Logical. Should \code{\link[turboEM:turboem]{turboEM::turboem()}} be used for EM algorithm? Default is TRUE.}

\item{calculate_confidence_interval}{Logical. Should confidence intervals be calculated? Default is FALSE.}
}
\value{

}
\description{
This function evaluate the presence (calls) of individual mutations from a predefined list.
TODO: Add describeIn instead
}
\seealso{
\code{\link[=call_cancer]{call_cancer()}}, \code{\link[=train_dreams_model]{train_dreams_model()}}
}
