% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_mutations_indels.R
\name{call_mutations_indels}
\alias{call_mutations_indels}
\title{Call mutations from read positions}
\usage{
call_mutations_indels(
  mutations_df,
  read_positions_df,
  read_positions_df_indels,
  model,
  model_indels,
  beta,
  beta_indels,
  alpha = 0.05,
  use_turboem = TRUE,
  calculate_confidence_intervals = FALSE,
  batch_size = NULL
)
}
\arguments{
\item{mutations_df}{A data frame containing the mutation information.}

\item{read_positions_df}{A data frame containing read position information for base substitution mutations.}

\item{read_positions_df_indels}{A data frame containing read position information for INDELs.}

\item{model}{The model used for calling base substitution mutations.}

\item{model_indels}{The model used for calling INDEL mutations.}

\item{beta}{Coefficients or parameters for the base substitution model.}

\item{beta_indels}{Coefficients or parameters for the INDEL model.}

\item{alpha}{Significance level for mutation detection; default is 0.05.}

\item{use_turboem}{Logical; if TRUE, use the TurboEM algorithm for faster computation.}

\item{calculate_confidence_intervals}{Logical; if TRUE, calculate confidence intervals for the estimated tumor fraction.}

\item{batch_size}{Number of positions to process at a time; if NULL, all positions are processed together.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with information about the individual mutation calls, including:
\describe{
\item{chr, genomic_pos}{The genomic position of the mutation.}
\item{ref, alt}{The reference and alternative allele.}
\item{EM_converged}{If the EM algorithm converged.}
\item{EM_steps, fpeval, objfeval}{Number of steps and function evaluations by the EM algorithm.}
\item{tf_est}{The estiamted tumor fraction (allele fraction).}
\item{tf_min, tf_max}{The confidence interval of \code{tf_est}.}
\item{exp_count}{The expected count of the alternative allele under the error (null) model.}
\item{count}{The count of the alternative allele.}
\item{coverage}{The coverage used by the model (only referenceredas with and alternative allele).}
\item{full_coverage}{The total coverage of the position (for reference).}
\item{obs_freq}{The observed frequency of the alternative allele.}
\item{ll_0, ll_A}{The value of the log-likelihood function under the null (tf=0) and alternative (tf>0) hypothesis.}
\item{Q_val, df, p_val}{The chisq test statistic, degrees of freedom and p-value of the statistical test.}
\item{mutation_detected}{Whether the mutation was detected at the supplied alpha level.}
}
}
\description{
@description This function evaluates the presence (calls) of
individual mutations, including insertions and deletions (INDELs), from a
predefined list. It processes given mutation and read position data frames
to call mutations based on the specified models and parameters.
}
\seealso{
\code{\link[=call_cancer]{call_cancer()}}, \code{\link[=train_dreams_model]{train_dreams_model()}}
}
\keyword{internal}
