% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{clean_insertions}
\alias{clean_insertions}
\title{Clean Insertions in CIGAR Sequence}
\usage{
clean_insertions(expanded_cigar)
}
\arguments{
\item{expanded_cigar}{A string representing the sequence output from
expand_cigar.}
}
\value{
A string of the CIGAR sequence after cleaning up the insertion
operations. For example, 'MMDDIMMII' would be to MMDDMI'.
}
\description{
This function adjusts the sequence of CIGAR operations by
cleaning up insertion operations. It removes any insertions that follow
deletions and consolidates consecutive insertions that follow any operation
other than deletions into a single insertion, while removing the operation
that precedes the insertions.
}
\keyword{internal}
