% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{combine_sequences}
\alias{combine_sequences}
\title{Combined CIGAR and MD tag}
\usage{
combine_sequences(cigar, md)
}
\arguments{
\item{cigar}{CIGAR string}

\item{md}{MD tag (\code{string})}
}
\value{
List of combined CIGAR and MD tag
}
\description{
Combined CIGAR and MD tag
}
\keyword{internal}
