% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{convert_to_cigar}
\alias{convert_to_cigar}
\title{Convert Expanded CIGAR String to Standard CIGAR Format}
\usage{
convert_to_cigar(cleaned_cigar)
}
\arguments{
\item{cleaned_cigar}{A character string representing the expanded CIGAR
string.}
}
\value{
A character string in standard CIGAR format.
}
\description{
This function takes an expanded CIGAR string (where each
operation is represented individually, e.g., "MMMMMMMDDDDDDDMIDD") and
converts it into standard CIGAR format (where consecutive operations are
counted, e.g., "7M7D1M1I2D").
}
\keyword{internal}
