% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_from_bam_indels.R
\name{correct_num}
\alias{correct_num}
\title{Correct Number Vector Based on CIGAR String}
\usage{
correct_num(cigar, numbers)
}
\arguments{
\item{cigar}{A character string representing a CIGAR sequence.}

\item{numbers}{A numeric vector that corresponds to the CIGAR string and needs correction.}
}
\value{
An integer vector with corrections applied based on the CIGAR string.
Deletions result in zeros inserted at the appropriate indices, and insertions are removed.
}
\description{
This function takes a CIGAR string and a vector of numbers and adjusts the numbers
based on the CIGAR operations. Deletions ('D') in the CIGAR string cause zeros to be inserted
into the number vector at the corresponding positions, and insertions ('I') cause the corresponding
numbers to be marked for removal. The final vector excludes these marked insertions.
}
