% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling.R
\name{correct_pos_idx_w_cigar}
\alias{correct_pos_idx_w_cigar}
\title{Title correct_pos_idx_w_cigar}
\usage{
correct_pos_idx_w_cigar(df)
}
\arguments{
\item{df}{dataframe_with_bam_data}
}
\value{
df
}
\description{
Title correct_pos_idx_w_cigar
}
\keyword{internal}
