% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cancer_indels.R
\name{dreams_cc_indels}
\alias{dreams_cc_indels}
\title{Cancer Detection in Indels Using DREAMS}
\usage{
dreams_cc_indels(
  mutations_df,
  bam_file_path,
  reference_path,
  model,
  alpha = 0.05,
  calculate_confidence_intervals = FALSE,
  use_turboem = TRUE
)
}
\arguments{
\item{mutations_df}{A dataframe containing the list of mutations to be analyzed.}

\item{bam_file_path}{Path to the BAM file containing sequencing data.}

\item{reference_path}{Path to the reference genome file, typically in FASTA format.}

\item{model}{The model to be used for cancer detection.}

\item{alpha}{Significance level for statistical testing, default is 0.05.}

\item{calculate_confidence_intervals}{Logical flag indicating whether to calculate confidence intervals, default is FALSE.}

\item{use_turboem}{Logical flag indicating whether to use the turboEM algorithm, default is TRUE.}
}
\value{
A \code{list()} with:
\itemize{
\item \strong{cancer_info} A \code{\link[=data.frame]{data.frame()}} with results for cancer calling across all mutations:
\describe{
\item{tf_est}{The estiamted tumor fraction (allele fraction).}
\item{tf_min, tf_max}{The confidence interval of \code{tf_est}.}
\item{r_est, est_mutations_present}{The estiamted fraction/number of candidate mutations present in the sample.}
\item{r_min, r_max}{The confidence interval of \code{r_est}.}
\item{mutations_tested}{Number of candidate mutations tested.}
\item{total_coverage, total_count}{Total count and coverage across all mutations (only reference and alternative allele(s).}
\item{mutations_tested}{Number of candidate mutations tested.}
\item{EM_converged}{If the EM algorithm converged.}
\item{EM_steps, fpeval, objfeval}{Number of steps and function evaluations by the EM algorithm.}
\item{ll_0, ll_A}{The value of the log-likelihood function under the null (tf=0) and alternative (tf>0) hypothesis.}
\item{Q_val, df, p_val}{The chisq test statistic, degrees of freedom and p-value of the statistical test.}
\item{cancer_detected}{Whether cancer was detected at the supplied alpha level.}
}

\item \strong{mutation_info} A \code{\link[=data.frame]{data.frame()}} with information about the individual mutations:
\describe{
\item{chr, genomic_pos}{The genomic position of the mutation.}
\item{ref, alt}{The reference and alternative allele.}
\item{P_mut_is_present}{The estimated probability the mutation is present in the sample.}
\item{exp_count}{The expected count of the alternative allele under the error (null) model.}
\item{count}{The count of the alternative allele.}
\item{coverage}{The coverage used by the model (only referenceredas with and alternative allele).}
\item{obs_freq}{The observed frequency of the alternative allele.}
}

}
}
\description{
This function performs cancer detection based on indel information from a BAM file using a predefined list of mutations.
It processes the mutations data and uses the provided model to make cancer calls.
}
\seealso{
\code{\link[=call_mutations]{call_mutations()}}, \code{\link[=train_dreams_model]{train_dreams_model()}}
}
