% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_mutations.R
\name{dreams_vc}
\alias{dreams_vc}
\title{Call mutations in a bam-file}
\usage{
dreams_vc(
  mutations_df,
  bam_file_path,
  reference_path,
  model,
  beta,
  alpha = 0.05,
  prior_error_rates = NULL,
  use_turboem = TRUE,
  calculate_confidence_intervals = FALSE,
  batch_size = NULL
)
}
\arguments{
\item{mutations_df}{A \code{\link[=data.frame]{data.frame()}} with candidate mutations (SNVs) (chromosome, positions, reference and alternative)}

\item{bam_file_path}{Path to .BAM-file}

\item{reference_path}{Path to reference genome e.g. FASTA-file.}

\item{model}{A dreams model. See \code{\link[=train_dreams_model]{train_dreams_model()}}.}

\item{beta}{Down sampling parameter from \code{\link[=get_training_data]{get_training_data()}} for correcting the error-rates from the DREAMS model.}

\item{alpha}{Alpha-level used for testing and confidence intervals. Default is 0.05.}

\item{prior_error_rates}{Prior error rates}

\item{use_turboem}{Logical. Should \code{\link[turboEM:turboem]{turboEM::turboem()}} be used for EM algorithm? Default is TRUE.}

\item{calculate_confidence_intervals}{Logical. Should confidence intervals be calculated? Default is FALSE.}

\item{batch_size}{Number of positions to process at a time}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with information about the individual mutation calls, including:
\describe{
\item{chr, genomic_pos}{The genomic position of the mutation.}
\item{ref, alt}{The reference and alternative allele.}
\item{EM_converged}{If the EM algorithm converged.}
\item{EM_steps, fpeval, objfeval}{Number of steps and function evaluations by the EM algorithm.}
\item{tf_est}{The estiamted tumor fraction (allele fraction).}
\item{tf_min, tf_max}{The confidence interval of \code{tf_est}.}
\item{exp_count}{The expected count of the alternative allele under the error (null) model.}
\item{prior_error_rates}{Prior error rates.}
\item{count}{The count of the alternative allele.}
\item{coverage}{The coverage used by the model (only referenceredas with and alternative allele).}
\item{full_coverage}{The total coverage of the position (for reference).}
\item{obs_freq}{The observed frequency of the alternative allele.}
\item{ll_0, ll_A}{The value of the log-likelihood function under the null (tf=0) and alternative (tf>0) hypothesis.}
\item{Q_val, df, p_val}{The chisq test statistic, degrees of freedom and p-value of the statistical test.}
\item{mutation_detected}{Whether the mutation was detected at the supplied alpha level.}
}
}
\description{
This function evaluate the presence (calls) of individual mutations from a predefined list.
}
\seealso{
\code{\link[=call_mutations]{call_mutations()}}, \code{\link[=call_cancer]{call_cancer()}}, \code{\link[=train_dreams_model]{train_dreams_model()}}
}
