% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{expand_cigar}
\alias{expand_cigar}
\title{Expand CIGAR string}
\usage{
expand_cigar(cigar)
}
\arguments{
\item{cigar}{CIGAR string}
}
\value{
List of expanded CIGAR string
}
\description{
Expand CIGAR string
}
\keyword{internal}
