% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{expand_md}
\alias{expand_md}
\title{Expand MD tag}
\usage{
expand_md(md)
}
\arguments{
\item{md}{MD tag (\code{string})}
}
\value{
List of expanded MD tag
}
\description{
Expand MD tag
}
\keyword{internal}
