% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_from_bam_indels.R
\name{extract_features_from_bam_indels_negatives}
\alias{extract_features_from_bam_indels_negatives}
\title{Extract Features from BAM for Indels in Negative Samples}
\usage{
extract_features_from_bam_indels_negatives(bam_df, reference_path)
}
\arguments{
\item{bam_df}{A data frame representing BAM file reads.}

\item{reference_path}{Path to the reference used in feature extraction.}
}
\value{
A modified data frame where each entry is enriched with additional features, particularly
for indels in negative samples. The 'obs' column is added, which is set equal to the 'ref' column.
}
\description{
This function processes a BAM file data frame and a reference path to extract features
specifically for indels (insertions or deletions) in negative samples. It enhances the data frame
with observed values set to reference values.
}
\keyword{internal}
