% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{extract_indel_pos_len}
\alias{extract_indel_pos_len}
\title{Extract Indel Positions and lengths from BAM Data}
\usage{
extract_indel_pos_len(bam_df)
}
\arguments{
\item{bam_df}{A dataframe obtained from the \code{load_BAM} function, containing
BAM data including fields for CIGAR strings and positions.}
}
\value{
A dataframe detailing the genomic positions of indels. This includes
columns for the positions of indels and their respective
lengths and base composition.
}
\description{
This function processes a dataframe obtained from \code{load_BAM}cto
identify and extract the genomic positions of indels. It filters the
dataframe for reads that contain indels, as indicated by the CIGAR string,
and then calculates the precise genomic positions, length and bases of these indels.
}
\keyword{internal}
