% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{extract_lengths_deletions}
\alias{extract_lengths_deletions}
\title{Extract Lengths of Deletions from CIGAR Strings}
\usage{
extract_lengths_deletions(cigar)
}
\arguments{
\item{cigar}{A character vector of CIGAR strings, each representing the alignment
of a single read in a BAM file.}
}
\value{
A numeric vector where each element represents the total length of
deletions in the corresponding CIGAR string.
}
\description{
This function processes a given CIGAR string and extracts the lengths of all
deletion operations. It utilizes the \code{extract_lengths_indels} function
to parse the CIGAR string and isolate the lengths corresponding to deletions.
}
\seealso{
\code{\link{extract_lengths_indels}}
}
\keyword{internal}
