% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{extract_lengths_indels}
\alias{extract_lengths_indels}
\title{Extract lengths of insertions and deletions}
\usage{
extract_lengths_indels(cigar)
}
\arguments{
\item{cigar}{A character string representing a CIGAR string.}
}
\value{
A list containing two elements: \code{I_length} and \code{D_length}. Each
element is a vector of numbers (as character strings) that appear before
'I' and 'D' in the CIGAR string.
}
\description{
This function extracts numeric lengths that appear before 'I' and 'D' in a
CIGAR string.
}
\keyword{internal}
