% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{extract_lengths_insertions}
\alias{extract_lengths_insertions}
\title{Extract Lengths of Insertions from CIGAR Strings}
\usage{
extract_lengths_insertions(cigar)
}
\arguments{
\item{cigar}{A character vector of CIGAR strings, each representing the alignment
of a single read in a BAM file.}
}
\value{
A numeric vector where each element represents the total length of
insertions in the corresponding CIGAR string.
}
\description{
This function processes a given CIGAR string and extracts the lengths of all
insertion operations. It leverages the \code{extract_lengths_indels} function
to parse the CIGAR string and isolate the lengths corresponding to insertions.
}
\keyword{internal}
