% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling.R
\name{extract_mismatch_positions}
\alias{extract_mismatch_positions}
\title{Title Get mismatch positions}
\usage{
extract_mismatch_positions(bam_df)
}
\arguments{
\item{bam_df}{dataframe from load_BAM}
}
\value{
mismatch_positions
}
\description{
Title Get mismatch positions
}
\keyword{internal}
