% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{extract_mismatch_positions_indels}
\alias{extract_mismatch_positions_indels}
\title{Extract indels positions from BAM data}
\usage{
extract_mismatch_positions_indels(bam_df)
}
\arguments{
\item{bam_df}{A dataframe obtained from the \code{load_BAM} function.}
}
\value{
A dataframe of mismatch positions, which includes the genomic
positions of all identified indels across the reads.
}
\description{
This function filters a dataframe for reads that contain indels
when compared to the reference genome, as indicated by the cigar string in
the BAM data. It then extracts the genomic positions of these mismatches,
providing a detailed account of where the sequencing reads do not match the
reference sequence.
}
\keyword{internal}
