% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R, R/train_model_indels.R
\name{fit_model}
\alias{fit_model}
\title{Title}
\usage{
fit_model(
  features,
  labels,
  input_structure,
  lr,
  batch_size,
  epochs,
  min_delta,
  patience,
  validation_split,
  model_file_path,
  log_file_path = NULL
)

fit_model(
  features,
  labels,
  input_structure,
  lr,
  batch_size,
  epochs,
  min_delta,
  patience,
  validation_split,
  model_file_path,
  log_file_path = NULL
)
}
\arguments{
\item{features}{input features}

\item{labels}{input labels}

\item{input_structure}{input structure}

\item{lr}{learning rate}

\item{batch_size}{batch size}

\item{epochs}{number of training epochs}

\item{min_delta}{minimum delta for early stopping}

\item{patience}{patience for early stopping}

\item{validation_split}{validation split}

\item{model_file_path}{output model file path}

\item{log_file_path}{output log file path}
}
\value{
fitted model

fitted model
}
\description{
Title

Title
}
\keyword{internal}
