% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{generate_NN_structure}
\alias{generate_NN_structure}
\title{Title}
\usage{
generate_NN_structure(inputs, input_layer, layers, reg = 0)
}
\arguments{
\item{inputs}{input structure}

\item{input_layer}{input layer}

\item{layers}{layer sizes}

\item{reg}{regularization}
}
\value{
NN structure
}
\description{
Title
}
\keyword{internal}
