% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_indels.R
\name{generate_NN_structure_indels}
\alias{generate_NN_structure_indels}
\title{Generate Neural Network Structure for Indel Analysis}
\usage{
generate_NN_structure_indels(inputs, input_layer, layers, reg = 0)
}
\arguments{
\item{inputs}{The input structure for the neural network, usually from the data input layer.}

\item{input_layer}{The initial layer of the neural network to which additional layers are added.}

\item{layers}{A list or vector indicating the number of units in each hidden layer of the network.}

\item{reg}{Regularization parameter applied to each dense layer for preventing overfitting.
Defaults to 0.}
}
\value{
A keras model object representing the constructed neural network.
}
\description{
This function builds a neural network structure for indel analysis. It takes
an input layer and iteratively adds specified hidden layers, followed by an
output layer for classification.
}
\keyword{internal}
