% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{generate_training_samples}
\alias{generate_training_samples}
\title{Title}
\usage{
generate_training_samples(
  bam_path,
  reference_path,
  bed_include_path = NULL,
  factor = 1,
  positions_to_exclude_paths = NULL,
  mm_rate_max = 1
)
}
\arguments{
\item{bam_path}{Path to BAM file}

\item{reference_path}{Path to reference file}

\item{bed_include_path}{BED regions to include}

\item{factor}{ratio between negative and positive data}

\item{positions_to_exclude_paths}{positions to exclude from training}

\item{mm_rate_max}{maximum mismatch rate in position}
}
\value{
dataframe with training data for a bam file
}
\description{
Title
}
