% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{get_genomic_pos_indels}
\alias{get_genomic_pos_indels}
\title{Extract Indel Information from CIGAR String and Genomic Position}
\usage{
get_genomic_pos_indels(pos, cigar)
}
\arguments{
\item{pos}{An integer representing the starting genomic position of the CIGAR
string.}

\item{cigar}{A character string representing the CIGAR format.}
}
\value{
An integer vector of the genomic positions of the indels.
}
\description{
This function takes a genomic position and a CIGAR string as input and
extracts information about indels. It processes the CIGAR string to determine
the genomic positions.
}
