% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{get_indel_info}
\alias{get_indel_info}
\title{Extract Indel Information from a CIGAR String}
\usage{
get_indel_info(pos, cigar)
}
\arguments{
\item{pos}{Integer, the starting genomic position of the CIGAR string.}

\item{cigar}{String, the CIGAR string representing genomic alignments.}
}
\value{
A list containing three elements: \code{genomic_pos} ,
\code{indel_length}, and \code{indel_type} .
}
\description{
This function takes a genomic position and a CIGAR string as input and extracts
information about indels present in the CIGAR string.
It returns the genomic positions, lengths, and types of these indels.
}
\keyword{internal}
