% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{get_indel_length_with_zero}
\alias{get_indel_length_with_zero}
\title{Get Indel Length with Zero Default}
\usage{
get_indel_length_with_zero(pos, cigar)
}
\arguments{
\item{pos}{A numerical value or vector representing the position(s).}

\item{cigar}{A string or vector of strings representing the CIGAR data.}
}
\value{
A numerical value or vector representing the indel lengths.
If the indel length calculated by \code{get_indel_length} is \code{NULL} or 0,
this function returns 0 instead.
}
\description{
This function calculates the indel length based on the provided position (pos)
and cigar string (cigar). It wraps around the \code{get_indel_length} function and ensures
that a 0 is returned whenever the indel length is either \code{NULL} or 0.
}
