% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{get_indel_type}
\alias{get_indel_type}
\title{Extract Types of Indels from a CIGAR String}
\usage{
get_indel_type(pos, cigar)
}
\arguments{
\item{pos}{Integer, the starting genomic position for the CIGAR string.}

\item{cigar}{String, the CIGAR string representing genomic alignments.}
}
\value{
A character vector containing the types of indels (either 'I' for insertions or 'D' for deletions)
in the CIGAR string.
}
\description{
This function utilizes \code{get_indel_info} to determine the types
of indels in a given CIGAR string.
}
\keyword{internal}
