% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{get_minimum_position}
\alias{get_minimum_position}
\title{Get the Minimum Position from Insertions and Deletions}
\usage{
get_minimum_position(pos, cigar)
}
\arguments{
\item{pos}{Integer, the start position of the CIGAR string in the genome.}

\item{cigar}{String, the CIGAR string representing genomic alignments.}
}
\value{
An integer representing the minimum position among the insertion and deletion positions.
Returns \code{NA} if there are no insertions or deletions.
}
\description{
This function combines the positions of insertions ('I') and deletions ('D')
obtained from a CIGAR string and returns the minimum genomic position among them.
}
\examples{
# Assuming you have a pos and a cigar string
# pos <- 10
# cigar <- "8M1I4M2D"
# min_position <- get_minimum_position(pos, cigar)
# print(min_position)

}
