% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{get_positions_of_indels}
\alias{get_positions_of_indels}
\title{Get genomic positions of insertions and deletions}
\usage{
get_positions_of_indels(pos, cigar)
}
\arguments{
\item{pos}{Start position of CIGAR string in genome.}

\item{cigar}{CIGAR string.}
}
\value{
A list with two elements:
- \code{I_positions}: a numeric vector with the positions of 'I's.
- \code{D_positions}: a numeric vector with the positions of the first 'D'
in each sequence of 'D's.
}
\description{
This function takes a CIGAR string as input and returns the
genomic start positions of insertions and deletions. Deletions start on
their actual position, whereas insertions start positions are annotated on
the base prior to the insertion.
}
\keyword{internal}
