% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_positions_from_BAM.R
\name{get_read_positions_from_BAM}
\alias{get_read_positions_from_BAM}
\title{Get read positions from BAM-file}
\usage{
get_read_positions_from_BAM(
  bam_file_path,
  chr,
  genomic_pos,
  reference_path,
  batch_size = NULL
)
}
\arguments{
\item{bam_file_path}{Path to BAM-file.}

\item{chr, genomic_pos}{Vectors. Should specify the positions of interest (\code{chr} = Chromosome, \code{genomic_pos} = Position in chromosome)}

\item{reference_path}{Path to reference genome e.g. FASTA-file.}

\item{batch_size}{Number of positions to process at a time}
}
\value{
\link{numeric}. Value for the sample specific beta
}
\description{
This functions extracts information about read position that cover positions of interest in a BAM-file.
}
