% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_from_bam.R
\name{get_reference_seq}
\alias{get_reference_seq}
\title{Get reference sequence}
\usage{
get_reference_seq(chr, genomic_pos, buffer, reference_path)
}
\arguments{
\item{chr}{chromosome}

\item{genomic_pos}{genomic_position}

\item{buffer}{how many neighbors to include}

\item{reference_path}{reference genome fa}
}
\description{
Get reference sequence
}
\keyword{internal}
