% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{get_training_data}
\alias{get_training_data}
\title{Extract training data from BAM files}
\usage{
get_training_data(
  bam_paths,
  reference_path,
  bed_include_path = NULL,
  factor = 1,
  common_positions_to_exclude_paths = NULL,
  positions_to_exclude_paths = NULL,
  mm_rate_max = 1
)
}
\arguments{
\item{bam_paths}{Vector of strings. Paths to \code{.bam} files to extract training data from.}

\item{reference_path}{String. Path to reference genome fasta file.}

\item{bed_include_path}{String. Path to bed-file with regions to include.}

\item{factor}{Number between 0 and 1. Ratio between negative and positive data.}

\item{common_positions_to_exclude_paths}{Vector of strings. List of files with positions to exclude from all samples.}

\item{positions_to_exclude_paths}{Vector of strings. List of files with positions to exclude from training with length equal to number of samples.}

\item{mm_rate_max}{Number between 0 and 1. Maximum mismatch rate in position.}
}
\value{
\code{data.frame} with training data for a bam file
}
\description{
Extract training data from BAM files
}
\seealso{
\code{\link[=train_model]{train_model()}} Function for training model
}
