% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{get_training_data_from_bam}
\alias{get_training_data_from_bam}
\title{Title}
\usage{
get_training_data_from_bam(
  bam_path,
  reference_path,
  bed_include_path = NULL,
  factor = 1,
  positions_to_exclude_paths = NULL,
  mm_rate_max = 1,
  chr = NULL,
  n_reads = NULL,
  read_fraction = NULL,
  get_beta = NULL,
  batch_size
)
}
\arguments{
\item{bam_path}{Path to BAM file}

\item{reference_path}{Path to reference file}

\item{bed_include_path}{BED regions to include}

\item{factor}{ratio between negative and positive data}

\item{positions_to_exclude_paths}{positions to exclude from training}

\item{mm_rate_max}{maximum mismatch rate in position}
}
\value{
dataframe with training data for a bam file
}
\description{
Title
}
\keyword{internal}
