% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data_indels.R
\name{get_training_data_from_bam_indel}
\alias{get_training_data_from_bam_indel}
\title{Extract Training Data from BAM File (indels)}
\usage{
get_training_data_from_bam_indel(
  bam_path,
  reference_path,
  bed_include_path = NULL,
  factor = 1,
  positions_to_exclude_paths = NULL,
  mm_rate_max = 1
)
}
\arguments{
\item{bam_path}{Path to the BAM file.}

\item{reference_path}{Path to the reference genome file.}

\item{bed_include_path}{Optional; BED file defining regions to include in the analysis.}

\item{factor}{The ratio of negative to positive data in the output.}

\item{positions_to_exclude_paths}{Optional; paths to files defining positions to exclude from training.}

\item{mm_rate_max}{Maximum mismatch rate allowed in a position.}
}
\value{
A list with two elements: \code{data}, a \code{data.frame} containing the combined positive and negative training data,
and \code{info}, a \code{data.frame} containing metadata about the training set.
}
\description{
Extracts training data from a BAM file by integrating information from reference and BED files.
It processes genomic positions of indels, extracts features, filters based on mismatch rates,
and combines positive and negative samples to form the training dataset.
}
\keyword{internal}
