% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BAM.R
\name{hardclip_correct_umi_features}
\alias{hardclip_correct_umi_features}
\title{Hard clip correct UMI features (ce and cd)}
\usage{
hardclip_correct_umi_features(df)
}
\arguments{
\item{df}{data.frame converted from lists-of-lists (scanBam)}
}
\value{
data.frame with corrected UMI features
}
\description{
Hard clip correct UMI features (ce and cd)
}
\keyword{internal}
