% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_error_rates.R
\name{predict_error_rates}
\alias{predict_error_rates}
\title{Predict error rates for read positions}
\usage{
predict_error_rates(read_positions_df, model, beta)
}
\arguments{
\item{read_positions_df}{Dataframe with read positions}

\item{model}{model file}

\item{beta}{beta value}
}
\description{
Predict error rates for read positions
}
