% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_error_rates.R
\name{predict_error_rates_parallel}
\alias{predict_error_rates_parallel}
\title{Parallel predict error rates}
\usage{
predict_error_rates_parallel(
  mutations_df,
  bam_file_path,
  reference_path,
  model,
  beta = NULL,
  factor = NULL,
  bed_file = NULL,
  ncores = 1,
  log_file = NULL,
  mm_rate_max = 0.05
)
}
\arguments{
\item{mutations_df}{A \code{\link[=data.frame]{data.frame()}} with candidate mutations (SNVs) (chromosome, positions, reference and alternative)}

\item{bam_file_path}{Path to .BAM-file}

\item{model}{A dreams model. See \code{\link[=train_dreams_model]{train_dreams_model()}}.}

\item{beta}{Down sampling parameter from \code{\link[=get_training_data]{get_training_data()}} for correcting the error-rates from the DREAMS model.}
}
\description{
This function evaluate the presence (calls) of individual mutations from a predefined list.
}
