% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{prepare_input_layer}
\alias{prepare_input_layer}
\title{Title}
\usage{
prepare_input_layer(training_data_features, ctx3_embed_dim)
}
\arguments{
\item{training_data_features}{training data features}

\item{ctx3_embed_dim}{trinucleotide context embedding dimensions}
}
\value{
input layer
}
\description{
Title
}
\keyword{internal}
