% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_indels.R
\name{prepare_input_layer_indels}
\alias{prepare_input_layer_indels}
\title{Prepare Input Layer for Indel Analysis in Neural Network}
\usage{
prepare_input_layer_indels(training_data_features, ctx3_embed_dim)
}
\arguments{
\item{training_data_features}{A data frame or similar structure containing
features of the training data.}

\item{ctx3_embed_dim}{Integer specifying the dimension of the embedding for
trinucleotide context sequences.}
}
\value{
A list containing two elements: \code{inputs}, the raw input from the
dataset, and \code{input_layer}, the prepared input layer for the neural network
model.
}
\description{
This function processes training data features and context embedding
dimensions to prepare an input layer suitable for a neural network model in
indel analysis.
}
\keyword{internal}
