% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{prepare_training_data}
\alias{prepare_training_data}
\title{Title}
\usage{
prepare_training_data(training_data, model_features)
}
\arguments{
\item{training_data}{training data}

\item{model_features}{selected features}
}
\value{
prepared training data
}
\description{
Title
}
