% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mutations_df_indels.R
\name{read_mutations_df}
\alias{read_mutations_df}
\title{Read and Modify Mutations Data Frame}
\usage{
read_mutations_df(mutations_df)
}
\arguments{
\item{mutations_df}{A data frame representing mutations. Each row should contain the fields 'ref', 'alt', and 'genomic_pos'.}
}
\value{
A modified data frame with updated mutation information.
}
\description{
This function processes a data frame containing mutation information. It modifies the 'ref', 'alt', and 'genomic_pos' columns based on certain conditions.
Specifically, if the 'ref' field contains more than one character, the function retains only the second character, marks the mutation as a deletion (by setting 'alt' to "D"),
and increments the 'genomic_pos' by 1. If the 'alt' field contains more than one character, the mutation is marked as an insertion (by setting 'alt' to "I").
}
\keyword{internal}
