% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BAM.R
\name{remove_hardclips}
\alias{remove_hardclips}
\title{Remove Hard Clipping from CIGAR Strings in a DataFrame}
\usage{
remove_hardclips(df)
}
\arguments{
\item{df}{A dataframe with a column named 'cigar', which contains CIGAR strings.}
}
\value{
A dataframe identical to the input, except the CIGAR strings in the
'cigar' column are modified to exclude initial and terminal hard clipping
and deletions.
}
\description{
This function processes a dataframe containing CIGAR strings
in one of its columns. It removes the hard clipping ('H') and deletions ('D')
at the beginning and end of each CIGAR string. Hard clipping and deletions
do not contribute to the alignment sequence, and their removal simplifies
further processing of these strings.
}
\keyword{internal}
