% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_from_bam_indels.R
\name{remove_insertions}
\alias{remove_insertions}
\title{Remove Insertions from Observed Sequence Based on CIGAR String}
\usage{
remove_insertions(cigar, obs_sequence)
}
\arguments{
\item{cigar}{A CIGAR string representing the alignment of an observed sequence
to a reference sequence.}

\item{obs_sequence}{The observed sequence (string) that is to be processed based
on the CIGAR string.}
}
\value{
A string representing the observed sequence with insertions removed.
}
\description{
This function processes an observed sequence based on a given CIGAR string,
removing any insertions (denoted as 'I' in the CIGAR string) from the observed sequence.
It first expands the CIGAR string to match the length of the observed sequence
and then iteratively checks each character. If a character in the expanded CIGAR string
is not an insertion ('I'), the corresponding character from the observed sequence
is retained in the result.
}
\keyword{internal}
