% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_softclips.R
\name{remove_softclips}
\alias{remove_softclips}
\title{Remove softclips}
\usage{
remove_softclips(df)
}
\arguments{
\item{df}{A dataframe from load_BAM}
}
\value{
A new dataframe with softclipped bases removed
}
\description{
Remove softclips
}
\keyword{internal}
