% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling.R
\name{sample_negative_read_positions}
\alias{sample_negative_read_positions}
\title{Title sample_negative_read_positions}
\usage{
sample_negative_read_positions(bam_df, n_samples)
}
\arguments{
\item{bam_df}{bam dataframe from load_BAM}

\item{n_samples}{number of samples}
}
\value{
negative_read_positions
}
\description{
Title sample_negative_read_positions
}
