% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{split_patient_data}
\alias{split_patient_data}
\title{Split Patient Data Into Training and Validation Sets}
\usage{
split_patient_data(training_data, patient_id_column, split_ratio = 0.8)
}
\arguments{
\item{training_data}{A list containing at least a \code{data} dataframe and an \code{info} object.
The \code{data} dataframe must contain a column specified by \code{patient_id_column}.}

\item{patient_id_column}{A string specifying the column name in \code{data} that contains the patient identifiers.
This column is used to split the data into training and validation sets.}

\item{split_ratio}{A numeric value between 0 and 1 indicating the proportion of the data to be used for training.
The default is 0.8, meaning 80\% of the data is used for training and 20\% for validation.}
}
\value{
A list with two elements: \code{train_data} and \code{validation_data}.
Each element is a list containing \code{data} and \code{info}, where \code{data} includes the respective split
of the original \code{data} dataframe, and \code{info} contains the unchanged \code{info} from \code{training_data}.
}
\description{
This function splits a dataset into training and validation sets based on a unique patient identifier.
It ensures that all records for a given patient are included entirely in either the training set or
the validation set, thereby preventing data leakage.
}
