% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BAM.R
\name{strand_correct_features}
\alias{strand_correct_features}
\title{Strand correct UMI features (ce and cd)}
\usage{
strand_correct_features(raw_bam_df)
}
\arguments{
\item{raw_bam_df}{data.frame converted from lists-of-lists (scanBam)}
}
\value{
strand corrected bam
}
\description{
Strand correct UMI features (ce and cd)
}
