% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BAM.R
\name{strand_correct_umi_features}
\alias{strand_correct_umi_features}
\title{Strand correct UMI features (ce and cd)}
\usage{
strand_correct_umi_features(df)
}
\arguments{
\item{df}{data.frame converted from lists-of-lists (scanBam)}
}
\value{
data.frame with corrected UMI features
}
\description{
This function corrects the orientation of Unique Molecular Identifier (UMI) features based on
the DNA strand from which they originate. It takes a data frame containing UMI features and adjusts them
if they are from the reverse strand, ensuring that the orientation of the UMI sequences is consistent with
their representation on the forward strand. This correction is necessary for accurate downstream analysis
of sequence data.
}
\keyword{internal}
