% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_dreams_model}
\alias{train_dreams_model}
\title{Train error model}
\usage{
train_dreams_model(
  training_data,
  layers,
  model_features,
  lr,
  batch_size,
  epochs,
  model_file_path = NULL,
  log_file_path = NULL,
  min_delta = 0,
  patience = 0,
  l2_reg = 0,
  validation_split = 0,
  ctx3_embed_dim = 3
)
}
\arguments{
\item{training_data}{\code{data.frame} Input training data (Generated from \code{\link[=get_training_data]{get_training_data()}})}

\item{layers}{Numeric vector. Number of nodes in each layer.}

\item{model_features}{Vector of feature names. Selected features for model training.}

\item{lr}{Numeric value between 0 and 1. Learning rate.}

\item{batch_size}{Integer. Batch size.}

\item{epochs}{Integer. Number of training epochs.}

\item{model_file_path}{String. Model output file path. Default is \code{NULL}.}

\item{log_file_path}{String. Path to model output log file. Default is \code{NULL}.}

\item{min_delta}{Numeric value between 0 and 1. Minimum delta for early stopping. Default is 0.}

\item{patience}{Integer. Patience when reaching minimum delta. Default is 0.}

\item{l2_reg}{Numeric value between 0 and 1. Level of L2 regularization per layer. Default is 0.}

\item{validation_split}{Numeric value between 0 and 1. Validation split ratio. Default is 0.}

\item{ctx3_embed_dim}{Integer. Number of dimensions to embed trinucleotide context to. Default is 3.}
}
\value{
Trained model in hdf5 format.
}
\description{
Train error model
}
\seealso{
\code{\link[=get_training_data]{get_training_data()}} Function for getting training data
}
\concept{Train model}
\keyword{model}
\keyword{training}
