% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Title}
\usage{
train_model(
  training_data,
  layers,
  model_features,
  lr,
  batch_size,
  epochs,
  model_file_path = NULL,
  log_file_path = NULL,
  decay = 0,
  min_delta = 0,
  patience = 0,
  l2_reg = 0,
  validation_split = 0,
  ctx3_embed_dim = 3
)
}
\arguments{
\item{training_data}{input training data}

\item{layers}{number of nodes in layers}

\item{model_features}{selected features}

\item{lr}{learning rate}

\item{batch_size}{batch size}

\item{epochs}{number of training epochs}

\item{model_file_path}{model output file path}

\item{log_file_path}{model output log file path}

\item{decay}{decay rate}

\item{min_delta}{minimum delta for early stopping}

\item{patience}{patient when reaching minimum delta}

\item{l2_reg}{level of L2 reg pr layer}

\item{validation_split}{validation split ratio}

\item{ctx3_embed_dim}{number of dimensions to embed trinucleotide context to}
}
\value{
trained model
}
\description{
Title
}
