% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{bed_to_granges}
\alias{bed_to_granges}
\title{Title}
\usage{
bed_to_granges(bed_path)
}
\arguments{
\item{bed_path}{path to bed-file}
}
\value{
granges object
}
\description{
Title
}
\keyword{internal}
