% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_cancer.R
\name{call_cancer}
\alias{call_cancer}
\title{Title}
\usage{
call_cancer(
  mutations_df,
  reads,
  model,
  beta,
  use_warp_speed = TRUE,
  calc_ci = FALSE
)
}
\arguments{
\item{mutations_df}{}
}
\value{

}
\description{
Title
}
