% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{combine_positions}
\alias{combine_positions}
\title{Combine Positions of Insertions and Deletions}
\usage{
combine_positions(pos, cigar)
}
\arguments{
\item{pos}{Integer, the start position of the CIGAR string in the genome.}

\item{cigar}{String, the CIGAR string representing genomic alignments.}
}
\value{
A numeric vector containing the combined positions of insertions and
deletions. Returns an empty vector if there are no insertions or deletions.
}
\description{
This function identifies and combines the positions of
insertions ('I') and deletions ('D'). The positions are returned as a
single vector, calculated based on the start position of the CIGAR string
in the genome.
}
\examples{
# Assuming you have a pos and a cigar string
# pos <- 10
# cigar <- "8M1I4M2D"
# combined_positions <- combine_positions(pos, cigar)

}
\keyword{internal}
