% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling.R
\name{correct_pos_idx_w_cigar}
\alias{correct_pos_idx_w_cigar}
\title{Correct positions index with CIGAR}
\usage{
correct_pos_idx_w_cigar(df)
}
\arguments{
\item{df}{dataframe_with_bam_data}
}
\value{
df
}
\description{
This function adjusts read positions in a dataframe based on CIGAR string information from a BAM file.
It processes the CIGAR string to identify insertions and deletions, which are then used to correct the read positions.
This is important for accurate read mapping and subsequent analyses.
}
\keyword{internal}
