% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling_indels.R
\name{expand_cigar}
\alias{expand_cigar}
\title{Expand CIGAR String into Sequence of Operations}
\usage{
expand_cigar(cigar)
}
\arguments{
\item{cigar}{A character string representing the CIGAR operations from an
alignment.}
}
\value{
A character string with each CIGAR operation expanded to show the
full sequence of alignment operations. For example, '2M1I' would expand to
'MMI'.
}
\description{
This function expands a condensed CIGAR string into a sequence
of operations.
}
\keyword{internal}
