% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_from_bam.R
\name{extract_features_from_bam}
\alias{extract_features_from_bam}
\title{Extract Features from BAM Data}
\usage{
extract_features_from_bam(
  bam_df,
  reference_path,
  add_umi_features = all(c("cd", "ce") \%in\% colnames(bam_df))
)
}
\arguments{
\item{bam_df}{A DataFrame originating from 'load_BAM' and processed by extract_mismatch_positions' containing alignment data.}

\item{reference_path}{Path to the reference genome file in FASTA format.}

\item{add_umi_features}{Check if umi information is available.}
}
\value{
A DataFrame with extracted features, including read positions and possibly UMI data.
}
\description{
Extract Features from BAM Data
}
\examples{
# Assuming 'bam_data' is a DataFrame obtained from 'load_BAM'
bam_data_processed <- extract_mismatch_positions(bam_data)
features <- extract_features_from_bam(bam_data_processed, "/path/to/reference.fasta")

}
\keyword{internal}
