% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_features_from_bam.R
\name{extract_features_from_bam}
\alias{extract_features_from_bam}
\title{Title}
\usage{
extract_features_from_bam(bam_df, reference_path)
}
\arguments{
\item{bam_df}{dataframe from load_BAM}

\item{reference_path}{reference genome path}
}
\value{
dataframe with read positions
}
\description{
Title
}
