% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling.R
\name{get_match_genomic_pos_list}
\alias{get_match_genomic_pos_list}
\title{Title get_match_genomic_pos_list}
\usage{
get_match_genomic_pos_list(pos, cigar, MDtag)
}
\arguments{
\item{pos}{position_list}

\item{cigar}{cigar string list}

\item{MDtag}{mdtag list}
}
\value{
match_positions
}
\description{
Title get_match_genomic_pos_list
}
