% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_position_handling.R
\name{get_mismatch_genomic_pos_list}
\alias{get_mismatch_genomic_pos_list}
\title{Title get_mismatch_genomic_pos_list}
\usage{
get_mismatch_genomic_pos_list(pos, MDtag)
}
\arguments{
\item{pos}{Genomic position}

\item{MDtag}{MDtag from BAM column}
}
\value{
genomic positions
}
\description{
Title get_mismatch_genomic_pos_list
}
