% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_positions_from_BAM.R
\name{get_read_positions_from_BAM}
\alias{get_read_positions_from_BAM}
\title{Get read positions from BAM-file}
\usage{
get_read_positions_from_BAM(bam_file_path, chr, genomic_pos, reference_path)
}
\arguments{
\item{bam_file_path}{Path to BAM-file.}

\item{chr, genomic_pos}{Vectors. Should specify the positions of interest (\code{chr} = Chromosome, \code{genomic_pos} = Position in chromosome)}

\item{reference_path}{Path to reference genome e.g. FASTA-file.}
}
\value{
\code{\link[=data.frame]{data.frame()}}. Each line describes a position in a read.
}
\description{
This functions extracts information about read position that cover positions of interest in a BAM-file.
}
