% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_training_data.R
\name{get_training_data}
\alias{get_training_data}
\title{Title}
\usage{
get_training_data(
  bam_paths,
  reference_path,
  bed_include_path = NULL,
  factor = 1,
  common_positions_to_exclude_paths = NULL,
  positions_to_exclude_paths = NULL,
  mm_rate_max = 1
)
}
\arguments{
\item{bam_paths}{Paths to BAM files}

\item{reference_path}{Path to reference file}

\item{bed_include_path}{BED regions to include}

\item{factor}{ratio between negative and positive data}

\item{common_positions_to_exclude_paths}{positions to exclude from all samples}

\item{positions_to_exclude_paths}{positions to exclude from training with length equal to number of samples}

\item{mm_rate_max}{maximum mismatch rate in position}
}
\value{
dataframe with training data for a bam file
}
\description{
Title
}
