% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BAM.R
\name{load_BAM}
\alias{load_BAM}
\title{Load .BAM file into R as tibble}
\usage{
load_BAM(BamPath, chr = NULL, pos = NULL)
}
\arguments{
\item{BamPath}{Path to .BAM-file}

\item{chr}{Vector of chromosomes of interest (strings)}

\item{pos}{Vector of positions of interest (numeric)}
}
\value{
.BAM file in tibble format
}
\description{
Load .BAM file into R as tibble
}
\keyword{internal}
