% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_error_rates.R
\name{predict_error_rates_parallel}
\alias{predict_error_rates_parallel}
\title{Parallel predict error rates}
\usage{
predict_error_rates_parallel(
  mutations_df,
  bam_file_path,
  reference_path,
  model,
  prior_error_rates = NULL,
  beta = NULL,
  ncores = 1,
  log_file = NULL,
  batch_size = NULL
)
}
\arguments{
\item{mutations_df}{A \code{\link[=data.frame]{data.frame()}} with candidate mutations (SNVs) (chromosome, positions, reference and alternative)}

\item{bam_file_path}{Path to .BAM-file}

\item{reference_path}{Path to reference genome e.g. FASTA-file.}

\item{model}{A dreams model. See \code{\link[=train_dreams_model]{train_dreams_model()}}.}

\item{beta}{Down sampling parameter from \code{\link[=get_training_data]{get_training_data()}} for correcting the error-rates from the DREAMS model.}

\item{ncores}{Number of processing cores}

\item{log_file}{write log-file to this path}

\item{batch_size}{Number of positions to process at a time}
}
\description{
This function evaluate the presence (calls) of individual mutations from a predefined list.
}
