% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BAM.R
\name{remove_softclips}
\alias{remove_softclips}
\title{Remove softclips}
\usage{
remove_softclips(df)
}
\arguments{
\item{df}{A dataframe from load_BAM}
}
\value{
A new dataframe with softclipped bases removed
}
\description{
This function processes a dataframe obtained from a BAM file and removes soft clipped bases from the sequences.
Soft clipping (indicated by 'S' in the CIGAR string) refers to bases that are present in the sequencing read but not used
in the alignment to the reference genome. This function trims these soft clipped bases from the sequence (\code{seq}) and quality
scores (\code{qual}), and also adjusts the CIGAR string accordingly. If UMI features are present, it will also trim these.
}
\keyword{internal}
