% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_BAM.R
\name{strand_correct_umi_features}
\alias{strand_correct_umi_features}
\title{Strand correct UMI features (ce and cd)}
\usage{
strand_correct_umi_features(df)
}
\arguments{
\item{df}{data.frame converted from lists-of-lists (scanBam)}
}
\value{
data.frame with corrected UMI features
}
\description{
Strand correct UMI features (ce and cd)
}
\keyword{internal}
